refactor: split src/bundle.rs (986) and src/recipe.rs (540) into modules (Refs PMAT-043)
Closes the 500-line-cap falsification that PMAT-042 overrode under
PMAT-043's accountability ticket. Every file in the new layout is ≤500.

Before                   After (all ≤500)
------                   -----------------
src/bundle.rs (986)  →   src/bundle/mod.rs         (37 — re-exports only)
                         src/bundle/v1.rs          (359 — APR v1 types + tests)
                         src/bundle/v2.rs          (449 — APR v2 types + impl)
                         src/bundle/v2_tests.rs    (155 — APR v2 tests, included
                                                    from v2.rs via #[path])
src/recipe.rs (540)  →   src/recipe/mod.rs         (25 — re-exports only)
                         src/recipe/context.rs     (262 — RecipeContext, Metadata,
                                                    MetricValue + impls)
                         src/recipe/testdata.rs    (41 — hash_name_to_seed,
                                                    generate_test_data,
                                                    generate_model_payload)
                         src/recipe/tests.rs       (227 — tests, included from
                                                    mod.rs via #[path])

Public API preserved exactly. `mod.rs` in each directory does
`pub use self::<submod>::*;` so every pre-existing
`use apr_cookbook::bundle::{...}` / `use apr_cookbook::recipe::{...}`
import continues to resolve.

Non-mechanical tweaks made during the split:
- bundle/v1.rs: `APR_MAGIC` and `MIN_HEADER_SIZE` changed from module-private
  to `pub(super) const`. v1 tests in the same file reference them. v2.rs
  does not use either (hardcodes `b"APR2"` and its own 64-byte v2 header).
- bundle/v2_tests.rs: explicit `use crate::bundle::v1::ModelBundle` so
  `test_v2_rejects_v1_format` can construct a v1 bundle.
- bundle/v2.rs: `#[cfg(test)] #[path = "v2_tests.rs"] mod tests;` so the
  tests live beside — not inside — v2.rs, keeping v2.rs at 449 lines.
- recipe/context.rs: explicit `use super::testdata::hash_name_to_seed`
  so `RecipeContext::new` / `verify_idempotency` can still reach it.
- recipe/tests.rs: `use crate::recipe::*` plus explicit `use std::time::Duration`
  (original module picked up Duration transitively through context.rs's imports).

src/lib.rs unchanged — `pub mod bundle;` / `pub mod recipe;` resolve to
the new `mod.rs` files via Cargo's directory-module rule.

Verification:
  wc -l src/bundle/*.rs src/recipe/*.rs      -> max 449 (all ≤500) ✓
  cargo test --all                            -> 128 passed, 0 failed
  cargo clippy --all-targets -- -D warnings   -> rc 0
  cargo fmt --all -- --check                  -> rc 0
